build(nix): `shell.nix` now depends on the output of cabal2nix
There were 3 main changes in this commit that revolve around cabal2nix. 1. `mad-props.nix` is generated by `cabal2nix` and is used by `defualt.nix` and `shell.nix`. 2. `defualt.nix` was added to work with `nix-build`. Currently it just passes off the environment that it creates to `shell.nix` but when/if it changes it will already exist. 3. `shell.nix` now depends on the output of `cabal2nix` to create a development environment for `nix-shell`.