fix(solve): removed `forall`s
I am not familiar with impredicative types but... removing the `forall`s in the solve function and its variants seems to have eliminated all type errors and now the project compiles and all of the examples work.

BREAKING CHANGE: As I said before I am not to familiar with impredicative types but it seems like this feature was added in GHC 9.2.1. Updated types for `solveT`, `solveAllT`, `solve`, and `solveAll`.